data: independent verification of the harmonised table
Three checks that go back to the ORIGINAL published files instead of comparing
generated artefacts against each other, which is all 09_verify can do.

90_roundtrip    every harmonised (effect, se) pair occurs in its source file.
                PASS, 54,076 of 54,076 across all 39 literatures.
91_distribution local parquet = local csv = live site = Zenodo deposit, to 1e-12,
                plus the invariants a user relies on. PASS.
92_pairing      was each effect/se pairing the unique one reproducing the file's
                own reported t, and by what margin? PASS, none beaten. 8 unique,
                9 tied with a near-collinear alternative, 22 not testable this way.

One real defect fixed: t_stat and precision were derived in float32 for the three
Stata-sourced literatures before widening, so they do not exactly equal effect/se
(494 and 667 rows, worst 5.5e-8 relative). 08_harmonise now widens first. The
shipped table is deliberately left alone -- regenerating would desynchronise the
site from the published 0.9.0-beta deposit over a difference no estimator can see.
It ships with 1.0.0; until then 91 asserts the deviation stays bounded by float32
epsilon and confined to those three literatures.

Both new checks failed falsely on the first run, twice from the same root cause:
float32 handling, and ^t$ matching the panel dimension T sitting beside N. Every
one was a defect in the test, not in the data.

Not wired into the pre-push hook: these are slow and 91 needs network.

+++ data_layer/08_harmonise.py
@@ -191,9 +191,14 @@ for proj in sorted(man):
     s=s.where(s>0)
     keep=e.notna()&s.notna()
     if keep.sum()<5: report[proj]=dict(included=False,reason=f"only {int(keep.sum())} usable rows"); continue
+    # float64 BEFORE deriving. Stata sources give float32, and t_stat/precision computed
+    # in float32 then stored as float64 do not reproduce effect/se exactly -- a user who
+    # recomputes gets a different number in the 8th digit. Immaterial for any estimator,
+    # but a derived column should be exactly derivable. Affected climate, euro,
+    # resource_curse; caught by 91_distribution.py.
     out=pd.DataFrame({"dataset":proj,
-                      "effect":e[keep].values,
-                      "se":s[keep].values})
+                      "effect":e[keep].astype("float64").values,
+                      "se":s[keep].astype("float64").values})
     out["t_stat"]=out["effect"]/out["se"]
     out["precision"]=1.0/out["se"]
     n_obs_was_log=False
